requirements.txt
aniso8601==7.0.0
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cryptography==39.0.1
defusedxml==0.7.1
Django==4.1.6
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-filter==22.1
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
flake8==6.0.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
promise==2.3
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==1.7.1
python-decouple==3.7
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
social-auth-app-django==5.0.0
sqlparse==0.4.3
style==1.1.0
text-unidecode==1.3
tomli==2.0.1
typing-extensions==4.4.0
update==0.0.1
urllib3==1.26.14
